requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-environ==0.9.0
psycopg2==2.9.3
python-dotenv==0.20.0
sqlparse==0.4.2

djangorestframework~=3.13.1
requests~=2.28.1